backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Azure SDK packages for full integration
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0
pyodbc==5.0.1
azure-cognitiveservices-language-textanalytics==5.2.0

# OpenAI Azure integration
openai==1.6.1

# Audio processing for real audio files
pydub==0.25.1
librosa==0.10.1
scipy==1.11.4

# Real-time features
websockets==12.0
requests==2.31.0

# Build tools to prevent compilation issues
setuptools>=65.0.0
wheel>=0.38.0